requirements.txt
python-dotenv>=1.0.0
structlog>=23.0.0
dependency-injector>=4.41.0
python-telegram-bot>=20.0
aiofiles>=23.0.0
aiohttp>=3.8.0
tiktoken>=0.5.0
redis>=4.5.0
numpy>=1.24.0
youtube-transcript-api>=0.6.0
PyYAML>=6.0
openai>=1.0.0
google-generativeai>=0.8.0
